switch to tailwind (cyg v2.1)
This is a massive overhaul!

Firstly, we are switching to tailwind. Bootstrap sucks and was a poor decision
originally. Tailwind makes developement easier with better helper classes and
better deduplication, especially when combined with postcss etc..

Secondly, we are introducing dark mode. Tailwind has great dark mode
support. Right now it is browser-based, but we may swap to a selector at
some point depending on feedback.

Thirdly, we are now using next/font, which is vastly superior to importing fonts
manually. It seems to be working as expected. No more pop-in!

Finally, there was some assorted cleanup done and a couple bugfixes added.
These probably shouldn't be in this commit but such is life.